Fixed the bug for not working on browser
