+++ package.json
@@ -25,7 +25,7 @@
   },
   "homepage": "https://dakc.github.io/merodb/",
   "devDependencies": {
-    "@babel/core": "^7.2.2",
+    "@babel/core": "^7.4.3",
     "@babel/preset-env": "^7.3.1",
     "babel-loader": "^8.0.5",
     "chai": "^4.2.0",
