+++ src/merodb.js
@@ -1,60 +1,6 @@
 'use strict';
-const fs = require('fs');
 const Core = require("./core");
 class MeroDB extends Core{
-    constructor(){
-        super();
-    }
-
-    /**
-     * write db contents to file
-     * @param {string} filePath - filepath to save db
-     */
-    save(filePath){
-        if(typeof filePath !== "string" || !filePath.trim()){
-            this.err = "filePath is not set";
-            return false;
-        }
-        this._writeOnFile(filePath);
-        return true;
-    }
-
-    /**
-     * write the contents of database
-     * on each insert,update and delete
-     * 
-     * @param {string} filePath  - filepath to save db
-     */
-    updateAlways(filePath){
-        if(typeof filePath !== "string" || !filePath.trim()){
-            this.err = "filePath is not set";
-            return false;
-        }
-        this.filePath = filePath;
-        this.isUpdateEveryTime = true;
-    }
-
-    /**
-     * load contents in a database from file
-     * @param {string} filePath 
-     */
-    loadFromFile(filePath){
-        try {
-            // check file existence
-            if (!fs.existsSync(filePath)) {
-                this.err = `database file ${filePath} does not exists.`;
-                return false;
-            }
-    
-            let data = fs.readFileSync(filePath);
-            this.clcn = JSON.parse(data);
-            return true;
-        } catch (error) {
-            this.err = error.message;
-        }
-        return false;
-    }
-
     /**
      * returns true if object is loaded in the database successfully
      * else returns false
